+++ example/App.vue
@@ -5,7 +5,6 @@
             <vue-rich-editor
                 :id="editorId"
                 :keepPasteFormat="keepPasteFormat"
-                :quillRegisterKeys="quillRegisterKeys"
                 useCustomImageHandler
                 useCustomImageLinkHandler
                 :disabled="editorIsDisabled"
@@ -37,7 +36,7 @@
             <h3>vue rich editor example2</h3>
             <vue-rich-editor
                 :id="editorId1"
-                :quillRegisterKeys="quillRegisterKeys1"
+                customLinkHref="http://realign.pw"
                 useCustomImageHandler
                 useCustomImageLinkHandler
                 :disabled="editorIsDisabled1"
@@ -67,7 +66,6 @@ export default {
             editorContent: '',
             setEditorDemo: '<h1>hahahah</h1>',
             editorIsDisabled: false,
-            quillRegisterKeys: ['inline', 'size', 'imageResize', 'imageLink'],
             linkPlaceholder: '请输入链接',
             editorFocusCache: null,
             editorHasFocusFlag: false,
@@ -76,7 +74,6 @@ export default {
             editorContent1: 'demo1<img src="https://public-bucket-realign.nos-eastchina1.126.net/image/normal/2018-12-30/img-1546104722969-415.png" width="536px" height="408px" />',
             setEditorDemo1: '<h2>hahahah</h2>',
             editorIsDisabled1: false,
-            quillRegisterKeys1: ['inline', 'size', 'imageResize', 'imageLink'],
             linkPlaceholder1: '请输入链接1'
         };
     },